# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Articles",
        "description": "Сервис статей",
        "contact": {
            "email": "av.pavlow@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        },
        {
            "url": "https://projects.dev/api/v1",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
        "/api/article": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Создание статьи",
                "description": "Сохранить",
                "operationId": "store",
                "requestBody": {
                    "description": "Поля",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Не переданы данные авторизации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Успешный ли запрос",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Доступ запрещен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ в api запрещен или приложение не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Успешный ли запрос",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Доступ запрещен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации."
                    },
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/article/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Получение статьи",
                "description": "Display the specified resource.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID статьи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Не переданы данные авторизации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Успешный ли запрос",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Доступ запрещен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ в api запрещен или приложение не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Успешный ли запрос",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Доступ запрещен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации."
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Удаление статьи",
                "description": "Remove the specified resource from storage.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID статьи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Articles"
                ],
                "summary": "Обновить статьи",
                "description": "Update the specified resource in storage.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID статьи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Поля",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Не переданы данные авторизации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Успешный ли запрос",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Доступ запрещен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации."
                    },
                    "403": {
                        "description": "Доступ в api запрещен или приложение не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Успешный ли запрос",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Доступ запрещен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/search": {
            "get": {
                "tags": [
                    "article"
                ],
                "summary": "List of articles",
                "description": "List of articles",
                "operationId": "articleList",
                "requestBody": {
                    "description": "Поля",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleSearchRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#Article"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/article/{id}/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Получение списка авторов статьи",
                "description": "Получение списка авторов статьи",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID статьи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthorResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Удаление адресата статьи",
                "description": "Remove the specified resource from storage.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID статьи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    },
                    {
                        "name": "author_id",
                        "in": "path",
                        "description": "Идентификатор автора",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthorResource"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Authors"
                ],
                "summary": "Обновление списка авторов статьи",
                "description": "Update the specified resource in storage.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID статьи",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Поля",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthorResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ArticleRequest": {
                "description": "Запрос создание статьи",
                "required": [
                    "categories",
                    "authors"
                ],
                "properties": {
                    "title": {
                        "description": "Заголовок",
                        "type": "string",
                        "example": ""
                    },
                    "source": {
                        "description": "Откуда взята статья",
                        "type": "string",
                        "example": ""
                    },
                    "description": {
                        "description": "Описание",
                        "type": "string",
                        "example": ""
                    },
                    "url": {
                        "description": "url",
                        "type": "string",
                        "example": ""
                    },
                    "image_url": {
                        "description": "Связанная картинка",
                        "type": "string",
                        "example": ""
                    },
                    "published_at": {
                        "description": "Дата публикации в формате Y-m-d\\TH:i:s.uP",
                        "type": "string",
                        "format": "date",
                        "example": "2021-10-31T09:15:57.572+03:00",
                        "nullable": true
                    },
                    "content": {
                        "description": "Контент",
                        "type": "string",
                        "example": ""
                    },
                    "type_name": {
                        "description": "Тип",
                        "type": "string",
                        "example": ""
                    },
                    "authors": {
                        "description": "Авторы статьи статьи",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Author"
                        }
                    },
                    "categories": {
                        "description": "Категории",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object"
            },
            "ArticleSearchRequest": {
                "description": "Запрос поиск статьи",
                "required": [
                    "category",
                    "authors"
                ],
                "properties": {
                    "type_name": {
                        "description": "Тип",
                        "type": "string",
                        "example": ""
                    },
                    "q": {
                        "description": "Текст",
                        "type": "string",
                        "example": ""
                    },
                    "from_date": {
                        "description": "Поиск от даты в формате Y-m-d\\TH:i:s.uP",
                        "type": "string",
                        "format": "date",
                        "example": "2021-10-31T09:15:57.572+03:00",
                        "nullable": true
                    },
                    "to_date": {
                        "description": "Поиск до даты в формате Y-m-d\\TH:i:s.uP",
                        "type": "string",
                        "format": "date",
                        "example": "2021-10-31T09:15:57.572+03:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}